Move test_resurrection to xplat file / spec out scenarios to run in xattr / doc meta modes
